backend/requirements.txt
fastapi
uvicorn
requests
pyyaml
apscheduler
